cleanup Table
