+++ src/test/scala/saferis/tests/TableQuerySpecs.scala
@@ -1,5 +1,6 @@
 package saferis.tests
 import saferis.*
+import saferis.DataManipulationLayer.*
 import zio.*
 import zio.test.*
